requirements.txt
# --- Web App Framework ---
streamlit==1.38.0
streamlit-audiorecorder==0.0.3

# --- Machine Learning & Deep Learning ---
tensorflow==2.17.0
scikit-learn==1.5.1
joblib>=1.3.2
pandas==2.2.2
numpy==1.26.4

# --- Computer Vision (New for Face Tab) ---
# Use 'headless' for servers/cloud (Streamlit Cloud), works locally too
opencv-python-headless>=4.10.0.84
Pillow>=10.3.0

# --- Audio Processing ---
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1
pydub==0.25.1
ffmpeg-python==0.2.0

# --- Visualization & Utilities ---
matplotlib==3.9.2
seaborn==0.13.2
nltk==3.9.1